item array defaulting to 0 - fixed
Fixed issue causing the array of items to default to index 0 when clicking the 'Generate' button and when changing modes.
Changing modes with an invalid length for that mode now defaults to index 0; changing modes with a valid length for that mode doesn't change the length value.
Tested. Confirmed working.